final commit

+++ src/br/com/alura/screenmatch/calculations/Stars.java
@@ -1,5 +1,5 @@
 package br.com.alura.screenmatch.calculations;
 
 public interface Stars {
-	int getStars();
+	double getStars();
 }
